+++ website/src/pages/demo/_constants/types.ts
@@ -17,6 +17,7 @@ export type DemoOption =
   | 'resetOnOperatorChange'
   | 'autoSelectField'
   | 'autoSelectOperator'
+  | 'autoSelectValue'
   | 'addRuleToNewGroups'
   | 'validateQuery'
   | 'independentCombinators'